requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiohttp
sqlmodel
cryptography
python-dotenv
authlib
jinja2
